smal fixes
