+++ Chroma/Chroma.csproj
@@ -46,15 +46,14 @@
     <Compile Include="HarmonyPatches\BombNoteController.cs" />
     <Compile Include="HarmonyPatches\SaberModelController.cs" />
     <Compile Include="HarmonyPatches\LightRotationEventEffect.cs" />
-    <Compile Include="HarmonyPatches\ScenesTransition\MultiplayerLevelScenesTransitionSetupDataSO.cs" />
-    <Compile Include="HarmonyPatches\ScenesTransition\SceneTransitionHelper.cs" />
     <Compile Include="HarmonyPatches\ScenesTransition\MissionLevelScenesTransitionSetupDataSO.cs" />
+    <Compile Include="HarmonyPatches\ScenesTransition\SceneTransitionHelper.cs" />
+    <Compile Include="HarmonyPatches\ScenesTransition\MultiplayerLevelScenesTransitionSetupDataSO.cs" />
     <Compile Include="HarmonyPatches\TrackLaneRingsRotationEffect.cs" />
     <Compile Include="LegacyLightHelper.cs" />
     <Compile Include="HarmonyPatches\BeatEffectSpawner.cs" />
     <Compile Include="HarmonyPatches\SaberBurnMarkArea.cs" />
     <Compile Include="HarmonyPatches\ScenesTransition\StandardLevelScenesTransitionSetupDataSO.cs" />
-    <Compile Include="HarmonyPatches\BeatmapObjectCallbackController.cs" />
     <Compile Include="LightColorManager.cs" />
     <Compile Include="Utils\Easings.cs" />
     <Compile Include="ChromaGradientController.cs" />
